feat: SignUpProcessTermsAndConditions
