+++ clients/giggle-client/src/components/SignUpForm/SignUpForm.tsx
@@ -1,31 +1,20 @@
 'use client'
 
 import LoginButton from '@/ui/Button/LoginButton'
-import { ChangeEvent, useCallback, useMemo, useState } from 'react'
+import { useCallback, useMemo } from 'react'
 import styled from 'styled-components'
 import { useSignUpStore } from '@/stores/SignUpStore'
-import SignUpFormEmail from './components/SignUpFormEmail'
-import SignUpFormPassword from './components/SignUpFormPassword'
-import SignUpFormUserInfo from './components/SignUpFormUserInfo'
-import SignUpFormTermsAndConditions from './components/SignUpFormTermsAndConditions/SignUpFormTermsAndConditions'
-import { useEffectOnce, useStartTyping } from 'react-use'
-import {
-  EmailSignUpActionParams,
-  emailSignUpAction,
-} from '../../../actions/signup'
+import { useEffectOnce } from 'react-use'
 import useSignUpRoute from './hooks/useSignUpRoute'
 import { match } from 'ts-pattern'
 import { z } from 'zod'
-import { emailSignInAction } from '../../../actions/login'
-import { useRouter } from 'next/navigation'
 import * as ReactAuth from 'next-auth/react'
-import SignUpFormEmailVerification from './components/SignUpFormEmailVerification'
 import { StepEnum } from './types'
-import LoadingOverlay from '../base/LoadingOverlay'
 import SignUpProcessEmailVerification from '../SignUpProcess/SignUpProcessEmailVerification'
 import SignUpProcessEmail from '../SignUpProcess/SignUpProcessEmail'
 import SignUpProcessPassword from '../SignUpProcess/SignUpProcessPassword'
 import SignUpProcessUserInfo from '../SignUpProcess/SignUpProcessUserInfo'
+import SignUpProcessTermsAndConditions from '../SignUpProcess/SignUpProcessTermsAndConditions'
 
 const TITLE_MESSAGE = `Sign up to start finding venues`
 
@@ -53,11 +42,7 @@ const Divider = styled.div`
 const StepSchema = z.number()
 
 export default function SignUpForm() {
-  const router = useRouter()
-  const { initializeStepRoute, increaseStepRoute, stepSearchParam } =
-    useSignUpRoute()
-
-  const [isLoading, setIsLoading] = useState(false)
+  const { initializeStepRoute, stepSearchParam } = useSignUpRoute()
 
   const step = useMemo<StepEnum | null>(() => {
     if (stepSearchParam === null) {
@@ -77,100 +62,18 @@ export default function SignUpForm() {
       .otherwise(() => null)
   }, [stepSearchParam])
 
-  const { setErrorMessage, errorMessage } = useSignUpStore((state) => ({
-    errorMessage: state.errorMessage,
-    setErrorMessage: state.setErrorMessage,
-  }))
-
-  const {
-    email,
-    password,
-    username,
-    termsAndConditions,
-    emailVerificationCode,
-  } = useSignUpStore((state) => ({
-    email: state.email,
-    password: state.password,
-    username: state.username,
-    termsAndConditions: state.termsAndConditions,
-    emailVerificationCode: state.emailVerificationCode,
-  }))
-  const {
-    setEmail,
-    setPassword,
-    setUsername,
-    setTermsAndConditions,
-    setEmailVerificationCode,
-  } = useSignUpStore((state) => ({
-    setEmail: state.setEmail,
-    setPassword: state.setPassword,
-    setUsername: state.setUsername,
-    setTermsAndConditions: state.setTermsAndConditions,
-    setEmailVerificationCode: state.setEmailVerificationCode,
-  }))
+  const { email, password, username, termsAndConditions } = useSignUpStore(
+    (state) => ({
+      email: state.email,
+      password: state.password,
+      username: state.username,
+      termsAndConditions: state.termsAndConditions,
+    })
+  )
 
   const onClickGoogleLoginButton = useCallback(async () => {
     await ReactAuth.signIn('google')
   }, [])
-
-  const handleSignUpFormTermsAndConditionsValidationError = useCallback(() => {
-    setErrorMessage('You have to check all mandatory terms')
-  }, [setErrorMessage])
-
-  const handleVerificationCodeInputChange = useCallback(
-    (e: ChangeEvent<HTMLInputElement>) => {
-      setEmailVerificationCode(e.target.value)
-    },
-    [setEmailVerificationCode]
-  )
-
-  const handleSignUpFormEmailValidationError = useCallback(
-    (reason: 'ZodValidation' | 'VerificationCodeNotCorrect') => {
-      match(reason)
-        .with('ZodValidation', () => {
-          setErrorMessage('Email verification code should be 6 digit number')
-        })
-        .with('VerificationCodeNotCorrect', () => {
-          setErrorMessage('Email verification code is not correct')
-        })
-        .exhaustive()
-    },
-    [setErrorMessage]
-  )
-
-  const handleSignUpSubmit = useCallback(async () => {
-    setIsLoading(true)
-    setErrorMessage('')
-    const needData: EmailSignUpActionParams = {
-      email,
-      password,
-      passwordConfirm: password,
-      verificationCode: emailVerificationCode,
-    }
-    try {
-      const response = await emailSignUpAction(needData)
-      if (response.isError) {
-        setErrorMessage(response.errorCode)
-        return
-      }
-      const signInResponse = await emailSignInAction({
-        ...needData,
-      })
-      if (signInResponse.isError) {
-        setErrorMessage(signInResponse.error)
-        return
-      }
-      await ReactAuth.signIn('credentials', {
-        redirect: true,
-        ...needData,
-      })
-      router.push('/')
-    } catch (e) {
-      console.error(e)
-    } finally {
-      setIsLoading(false)
-    }
-  }, [email, emailVerificationCode, password, router, setErrorMessage])
 
   useEffectOnce(() => {
     match(step)
@@ -210,16 +113,7 @@ export default function SignUpForm() {
         .with(null, () => <SignUpProcessEmail />)
         .with(1, () => <SignUpProcessPassword />)
         .with(2, () => <SignUpProcessUserInfo />)
-        .with(3, () => (
-          <SignUpFormTermsAndConditions
-            initialTermsAndConditions={termsAndConditions}
-            onUserCheckedTermsAndConditions={setTermsAndConditions}
-            onValidationError={
-              handleSignUpFormTermsAndConditionsValidationError
-            }
-            onSubmit={increaseStepRoute}
-          />
-        ))
+        .with(3, () => <SignUpProcessTermsAndConditions />)
         .with(4, () => <SignUpProcessEmailVerification />)
         .exhaustive()}
       <Divider />
@@ -227,8 +121,6 @@ export default function SignUpForm() {
         onClick={onClickGoogleLoginButton}
         fullWidth
       >{`${LOGIN_PRE_MESSAGE} Google`}</LoginButton>
-      {errorMessage}
-      <LoadingOverlay isLoading={isLoading} />
     </Wrapper>
   )
 }
